+++ bundle/src/main/java/com/adobe/acs/commons/granite/ui/components/impl/include/IncludeDecoratorFilterImpl.java
@@ -22,8 +22,12 @@ import org.apache.sling.api.SlingHttpServletRequest;
 import org.apache.sling.api.resource.Resource;
 import org.apache.sling.api.resource.ValueMap;
 import org.apache.sling.engine.EngineConstants;
+import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.ConfigurationPolicy;
+import org.osgi.service.component.annotations.Modified;
+import org.osgi.service.metatype.annotations.Designate;
+import org.osgi.service.metatype.annotations.ObjectClassDefinition;
 
 import javax.annotation.CheckForNull;
 import javax.servlet.Filter;
@@ -35,6 +39,9 @@ import javax.servlet.ServletException;
 
 
 import java.io.IOException;
+import java.util.Arrays;
+import java.util.List;
+import java.util.function.Predicate;
 
 @Component(
         service = Filter.class,
@@ -43,6 +50,7 @@ import java.io.IOException;
                 EngineConstants.SLING_FILTER_SCOPE+"="+ EngineConstants.FILTER_SCOPE_INCLUDE
         }
 )
+@Designate(ocd = IncludeDecoratorFilterImpl.Config.class)
 public class IncludeDecoratorFilterImpl implements Filter {
 
     static final String RESOURCE_TYPE = "acs-commons/granite/ui/components/include";
@@ -65,7 +73,28 @@ public class IncludeDecoratorFilterImpl implements Filter {
      * Prefix value parameters passed downwards
      */
     static final String PREFIX = "ACS_AEM_COMMONS_INCLUDE_PREFIX_";
-    
+
+
+    static final String REQ_ATTR_IGNORE_CHILDREN_RESOURCE_TYPE = "ACS_AEM_COMMONS_EXCLUDE_CHILDREN_RESOURCE_TYPE";
+
+    private List<String> resourceTypesIgnoreChildren;
+
+    @ObjectClassDefinition(
+            name = "ACS AEM Commons - IncludeDecoratorFilterImpl",
+            description = "Used to perform the namespacing / parameterization in the include context")
+    public @interface Config {
+        String[] resourceTypesIgnoreChildren() default {
+                "granite/ui/components/coral/foundation/form/multifield"
+        };
+    }
+
+    @Activate
+    @Modified
+    public void init(Config config) {
+        this.resourceTypesIgnoreChildren = Arrays.asList(config.resourceTypesIgnoreChildren());
+    }
+
+
     @Override
     public void init(FilterConfig filterConfig) throws ServletException {
         // no-op
@@ -77,16 +106,28 @@ public class IncludeDecoratorFilterImpl implements Filter {
         ValueMap parameters = ValueMap.EMPTY;
 
         if(servletRequest instanceof SlingHttpServletRequest){
-            
+
             SlingHttpServletRequest request = (SlingHttpServletRequest) servletRequest;
 
-            if(request.getResourceResolver().isResourceType(request.getResource(), RESOURCE_TYPE)){
+            Predicate<String> typeCheckFn = (resourceType) -> request.getResourceResolver().isResourceType(request.getResource(), resourceType);
+
+            if(typeCheckFn.test(RESOURCE_TYPE)){
                 performFilter(request, servletResponse, chain, parameters);
                 return;
+            }else if(resourceTypesIgnoreChildren.stream().anyMatch(typeCheckFn)){
+                boolean ignoreChildren = resourceTypesIgnoreChildren.stream().anyMatch(typeCheckFn);
+                if(ignoreChildren){
+                    request.setAttribute(REQ_ATTR_IGNORE_CHILDREN_RESOURCE_TYPE, request.getResource().getResourceType());
+                }
+                chain.doFilter(servletRequest, servletResponse);
+                if(ignoreChildren){
+                    request.removeAttribute(REQ_ATTR_IGNORE_CHILDREN_RESOURCE_TYPE);
+                }
+                return;
             }
-            
+
         }
-    
+
         chain.doFilter(servletRequest, servletResponse);
     }