+++ bundle/src/main/java/com/adobe/acs/commons/granite/ui/components/impl/include/NamespacedTransformedResourceProviderImpl.java
@@ -33,7 +33,7 @@ import org.osgi.service.metatype.annotations.ObjectClassDefinition;
 @Component(service = NamespacedTransformedResourceProvider.class)
 @Designate(ocd = NamespacedTransformedResourceProviderImpl.Config.class)
 public class NamespacedTransformedResourceProviderImpl implements NamespacedTransformedResourceProvider {
-    
+
     @ObjectClassDefinition(
             name = "ACS AEM Commons - NamespacedTransformedResourceProvider",
             description = "Transforms a resource underlying children with a namespace. Used for granite dialog snippets to be included in a granular way.")
@@ -44,21 +44,25 @@ public class NamespacedTransformedResourceProviderImpl implements NamespacedTran
         )
         String[] properties() default {"name", "fileNameParameter", "fileReferenceParameter"};
     }
-    
+
     @Reference
     private ExpressionResolver expressionResolver;
-    
+
     private String[] namespacedProperties;
-    
+
     @Activate
     @Modified
     public void init(Config config) {
         this.namespacedProperties = config.properties();
     }
-    
+
     @Override
     public Resource transformResourceWithNameSpacing(SlingHttpServletRequest request, Resource targetResource) {
-        return new NamespaceResourceWrapper(targetResource, expressionResolver, request, namespacedProperties);
+        return new NamespaceResourceWrapper(
+                targetResource,
+                expressionResolver,
+                request,
+                namespacedProperties);
     }
-    
+
 }